Add -XstartOnFirstThread for macOS client runs
