+++ src/main/java/net/neoforged/moddevgradle/internal/OperatingSystemDisambiguation.java
@@ -1,7 +1,6 @@
 package net.neoforged.moddevgradle.internal;
 
-
-import org.gradle.api.GradleException;
+import net.neoforged.moddevgradle.internal.utils.OperatingSystem;
 import org.gradle.api.attributes.AttributeDisambiguationRule;
 import org.gradle.api.attributes.MultipleCandidatesDetails;
 
@@ -11,18 +10,10 @@ import org.gradle.api.attributes.MultipleCandidatesDetails;
 public abstract class OperatingSystemDisambiguation implements AttributeDisambiguationRule<String> {
     @Override
     public void execute(MultipleCandidatesDetails<String> details) {
-        var osName = System.getProperty("os.name");
-        // The following matches the logic in Apache Commons Lang 3 SystemUtils
-        if (osName.startsWith("Linux") || osName.startsWith("LINUX")) {
-            osName = "linux";
-        } else if (osName.startsWith("Mac OS X")) {
-            osName = "osx";
-        } else if (osName.startsWith("Windows")) {
-            osName = "windows";
-        } else {
-            throw new GradleException("Unsupported operating system: " + osName);
-        }
-
-        details.closestMatch(osName);
+        details.closestMatch(switch (OperatingSystem.current()) {
+            case LINUX -> "linux";
+            case MACOS -> "osx";
+            case WINDOWS -> "windows";
+        });
     }
 }
